\name{is.url}
\alias{is.url}
\title{Check if the given object is an url string}
\usage{
  is.url(s)
}
\arguments{
  \item{s}{input object to check

  Returns:}
}
\value{
  TRUE/FALSE indicating whether the input string is a valid
  URL.
}
\description{
  Arguments:
}
\examples{
# is.url("http://aa.px")
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("statfi")
}
\keyword{utilities}

