% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statisticsfinland.R
\name{list_eurostat_files}
\alias{list_eurostat_files}
\title{List the open data files available from Eurostat}
\usage{
list_eurostat_files(...)
}
\arguments{
\item{...}{Arguments to be passed

Returns:}
}
\value{
table
}
\description{
Arguments:
}
\examples{
# df <- list_eurostat_files()
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("statfi")
}
\keyword{utilities}

