\name{list_statfi_files}
\alias{list_statfi_files}
\title{List the open data files available from Statistics Finland}
\usage{
  list_statfi_files(format = "px")
}
\arguments{
  \item{format}{"px", "csv", "xml" depending on the desired
  format of the data files

  Returns:}
}
\value{
  table
}
\description{
  Arguments:
}
\examples{
# df <- list_statfi_files()
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("statfi")
}
\keyword{utilities}

