\name{read.px}
\alias{read.px}
\title{read.px}
\usage{
read.px(filename, encoding = "latin1", na.strings = c("\\".\\"", "\\"..\\"",
  "\\"...\\"", "\\"....\\""))
}
\arguments{
  \item{filename}{filename}

  \item{encoding}{encoding}

  \item{na.strings}{na.strings}
}
\value{
px object
}
\description{
Fixed version of read.px from pxR package
}
\examples{
#
}
\author{
Contact: Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("statfi")
}
\keyword{utilities}

