\docType{package}
\name{statfi-package}
\alias{statfi}
\alias{statfi-package}
\title{R Tools for Statistics Finland}
\description{
Brief summary of the statfi package
}
\details{
\tabular{ll}{ Package: \tab statfi\cr Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2013\cr License: \tab FreeBSD\cr LazyLoad: \tab
yes\cr }

R Tools for Statistics Finland Open Data
}
\examples{
library(statfi)
}
\author{
Leo Lahti, Juuso Parkkinen and Joona Lehtomaki
\email{louhos@googlegroups.com}
}
\references{
See citation("statfi")
\url{http://github.com/ropengov/statfi}
}
\keyword{package}

